requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
boto3>=1.26.0
pydantic>=1.9.0
pydantic-settings>=2.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.5
reportlab>=3.6.0
pandas>=1.3.0
pytest>=6.2.5
pytest-cov>=2.12.0
python-dotenv>=0.19.0
websockets>=10.0
